requirements.txt
# Flask 框架和相关扩展
Flask==3.0.0  # Web 应用框架
gunicorn==21.2.0  # WSGI HTTP 服务器

# Celery 和 Redis 相关库
celery==5.3.4  # 异步任务队列
redis==5.0.1  # Redis 客户端

# 数据库相关库
psycopg2-binary==2.9.9  # PostgreSQL 数据库驱动

# YouTube 视频下载库
pytube==15.0.0  # 下载 YouTube 视频

# AWS S3 相关库
boto3==1.34.43  # AWS SDK for Python

# 其他常用库
python-dotenv==1.0.0  # 环境变量管理
flask-httpauth==4.8.0  # 基于 HTTP 的认证库

# 日志记录库
loguru==0.7.2  # 更强大的日志记录库（可选）